Added app_name + Removed unused fields in models + A lot of refactoring
