+++ src/feature_selection/models.py
@@ -117,8 +117,6 @@ class FSExperiment(models.Model):
     # AWS-EMR fields
     app_name = models.CharField(max_length=100, null=True, blank=True)  # Spark app name to get the results
     emr_job_id = models.CharField(max_length=100, null=True, blank=True)  # Job ID in the Spark cluster
-    molecules_path = models.CharField(max_length=200, null=True, blank=True)  # Path to the molecules shared folder
-    clinical_path = models.CharField(max_length=200, null=True, blank=True)  # Path to the molecules shared folder
 
     def get_all_sources(self) -> List[Optional['api_service.ExperimentSource']]:
         """Returns a list with all the sources."""
